+++ src/immutable/hashes.ts
@@ -1,39 +1,47 @@
+import { Varint } from '../core.js';
 import { Base58 } from '../internal/encoding.js';
+import type { CIDLike } from './cid.js';
 
-/**
- * Hashing algorithms available to the SDK.
- *
- * TODO: Replace with a Registry.
- */
+/** Hashing algorithms available to the SDK. */
 export enum HashAlgorithm {
   /** Type code for SHA-256. */
   SHA256 = 0,
 }
 
-/** TODO: Make like other encoding classes. */
+/** A parser for the hash part of the `ContentIdentifier` for immutable content. */
 export class Hash {
+  /** The full bytes. */
+  public bytes: Uint8Array;
+
   constructor(
-    /** The hash algorithm identifier byte of the hash. */
-    readonly algorithm: HashAlgorithm,
-    /** The hash bytes, minus the first algorithm identifier byte. */
-    readonly value: Uint8Array,
-  ) {}
-
-  /**
-   * Returns the full hash bytes, including the first algorithm identifier byte, as a new
-   * `Uint8Array`.
-   */
-  toBytes() {
-    return new Uint8Array([this.algorithm, ...this.value]);
+    /** A valid hash value. When a string is used, it must be base58 encoded. */
+    value: CIDLike,
+  ) {
+    this.bytes =
+      value instanceof Hash
+        ? value.bytes
+        : typeof value === 'string'
+          ? Base58.decode(value)
+          : new Uint8Array(value);
   }
 
-  /** Returns the hash encoded into base58. */
-  toBase58(): string {
-    return Base58.encode(this.toBytes());
+  /** The algorithm integer. */
+  get algorithm() {
+    return new Varint(this.bytes);
   }
 
-  /** Returns the hash encoded into base58. */
-  toString(): string {
+  /** The value as bytes. */
+  get value() {
+    return this.bytes.subarray(this.algorithm.encodingLength);
+  }
+
+  /** Gets a version encoded as human-readable base58 string. */
+  toBase58() {
+    return Base58.encode(this.bytes);
+  }
+
+  /** Alias for `toBase58`. */
+  toString() {
     return this.toBase58();
   }
 }
@@ -48,7 +56,7 @@ export class Hash {
 export async function hash(alg: HashAlgorithm, payload: BufferSource): Promise<Hash> {
   switch (alg) {
     case HashAlgorithm.SHA256:
-      return new Hash(alg, new Uint8Array(await sha256(payload)));
+      return new Hash([alg, ...new Uint8Array(await sha256(payload))]);
     default:
       throw new TypeError('Unsupported algorithm');
   }