+++ src/immutable/schema.ts
@@ -1,9 +1,8 @@
-import { decodeWithCodec } from '../codec/codecs.js';
 import { File } from '../file/file.js';
-import type { IdentifierSchema } from '../identifiers/identifiers.js';
+import type { ContentIdentifierScheme } from '../identifiers/identifiers.js';
 import { validateCID } from './cid.js';
 
-/** {@linkcode IdentifierSchema} for immutable content-addressed files. */
+/** {@linkcode ContentIdentifierScheme} for immutable content-addressed files. */
 export const Immutable = {
   /** Immutable data uses the identifier type `1`. */
   key: 1 as const,
@@ -16,12 +15,10 @@ export const Immutable = {
    * @param instanceID The instance for codec resolution.
    */
   async parse(cid, content, instanceID?: string) {
-    if (await validateCID(cid.value, content)) {
+    if (await validateCID(cid.rawValue, content)) {
       const file = new File(content);
-      return file.mediaType
-        ? // Will throw if content is malformed or unsupported
-          decodeWithCodec(file.payload, file.mediaType.value, instanceID)
-        : file.payload;
+      await file.getValue(instanceID); // validate
+      return file;
     }
   },
-} satisfies IdentifierSchema;
+} satisfies ContentIdentifierScheme<File>;